changes in maker and created storage

+++ ToiletRace/Managers/PooNodeCreator.swift
@@ -60,16 +60,16 @@ class PooNodeCreator {
     }
     
     static func createCustomizedBall(postion: SCNVector3?, item: BolusItem) -> SCNNode {
-        let geo = SCNSphere(radius: item.radius)
+        let geo = SCNSphere(radius: CGFloat(item.radius))
         geo.materials.insert(materialFromItem(item), at: 0)
         geo.materials.removeLast()
         let pooNode = SCNNode(geometry: geo)
         pooNode.position = postion ?? SCNVector3(0.0, 0.0, 0.0)
         pooNode.physicsBody = SCNPhysicsBody.dynamic()
-        pooNode.physicsBody?.restitution = item.restitution
+        pooNode.physicsBody?.restitution = CGFloat(item.restitution)
         pooNode.physicsBody?.contactTestBitMask = Collider.poo | Collider.obstacle | Collider.bounds
         pooNode.physicsBody?.categoryBitMask = Collider.poo
-        pooNode.physicsBody?.mass = item.mass
+        pooNode.physicsBody?.mass = CGFloat(item.mass)
         pooNode.name = SessionData.shared.selectedPlayer.name.rawValue
         return pooNode
     }
@@ -83,7 +83,7 @@ class PooNodeCreator {
         material.specular.contents = UIImage(named: "puffSpec")
         material.roughness.contents = 0.8
         material.displacement.contents = UIImage(named: "raisHei")
-        material.displacement.intensity = item.displacement
+        material.displacement.intensity = CGFloat(item.displacement)
         return material
     }
 }
